prime-minister/requirements.txt
click==6.2
python-docx==0.8.5
